+++ packages/logger/src/index.ts
@@ -377,3 +377,5 @@ export class optimizeEdgeWorkerLogger {
   }
 }
 export * from './kafkaStreamWorker';
+
+export * from './wasmLogFormatter';