+++ packages/react-table/src/index.ts
@@ -473,3 +473,5 @@ export function traverseCRDTMultiplayerHook(
   return state;
 }
 export * from './pivotEngine';
+
+export * from './inlineEditorState';