+++ packages/retry/src/index.ts
@@ -507,4 +507,5 @@ export class broadcastEventEmitterListener {
     }
   }
 }
-export * from './circuitState';
+
+export * from './circuitStateMachine';